Fixes applied
