+++ MovieApp/PresentationLayer/Presenters/HomeScreenPresenter.swift
@@ -1,5 +1,2 @@
-import Foundation
-
-class HomeScreenPresenter: HomeScreenPresenterProtocol {
-
+class HomeScreenPresenter {
 }
