--- MovieApp/PresentationLayer/Presenters/HomeScreenPresenterProtocol.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-protocol HomeScreenPresenterProtocol {
-    
-}
