+++ MovieApp/PresentationLayer/Views/ConstructViewsProtocol.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 protocol ConstructViewsProtocol {
     
     func createViews()
