--- MovieApp/PresentationLayer/Views/HomeScreen/HomeScreenDelegate.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-protocol HomeScreenDelegate {
-}
